requirements.txt
alembic==0.9.9
aniso8601==3.0.0
click==6.7
Flask==1.0.2
Flask-MySQL==1.4.0
flask-restplus==0.11.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
PyMySQL==0.8.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.8
Werkzeug==0.14.1
